fixed bug on product page
create new component for category page
