+++ src/entities/CatalogNavigation/ui/LinksList/LinksList.tsx
@@ -1,8 +1,6 @@
 import { RoutePath } from "shared/config/routeConfig/routeConfig"
 import { classNames } from "shared/lib/classNames/classNames"
 import { AppLink } from "shared/ui/AppLink/AppLink"
-import { useNavigate } from "react-router-dom"
-import productPlaceholder from "shared/assets/images/productPlaceholder.jpg"
 import { navigationCategory } from "../../model/types/list"
 import styles from "./LinksList.module.scss"
 
@@ -13,28 +11,20 @@ interface LinksListProps {
 
 export function LinksList({ data, className }: LinksListProps) {
     const { id, name, categories } = data
-    const navigate = useNavigate()
 
     return (
         <div className={classNames(styles.container, {}, [className])}>
             <AppLink to={RoutePath.category + id} className={styles.title}>
                 {name}
             </AppLink>
-            <div className={styles.listContainer}>
+            <div className={styles.list}>
                 {categories.map(item => {
-                    const { id, name, image } = item
-                    const categoryImage = image || productPlaceholder
+                    const { id, name } = item
 
                     return (
-                        <div
-                            onClick={() => navigate(RoutePath.sub_category + id)}
-                            className={styles.list}
-                        >
-                            <img src={categoryImage} alt="pic" className={styles.image} />
-                            <AppLink key={id} to={RoutePath.sub_category + id}>
-                                {name}
-                            </AppLink>
-                        </div>
+                        <AppLink key={id} to={RoutePath.sub_category + id}>
+                            {name}
+                        </AppLink>
                     )
                 })}
             </div>
