+++ src/pages/CategoryPage/ui/CategoryLinks/CategoryLinks.tsx
@@ -1,0 +1,43 @@
+import { RoutePath } from "shared/config/routeConfig/routeConfig"
+import { classNames } from "shared/lib/classNames/classNames"
+import { AppLink } from "shared/ui/AppLink/AppLink"
+import { useNavigate } from "react-router-dom"
+import { navigationCategory } from "entities/CatalogNavigation/model/types/list"
+import productPlaceholder from "shared/assets/images/productPlaceholder.jpg"
+import styles from "./CategoryLinks.module.scss"
+
+interface CategoryLinksProps {
+    data: navigationCategory
+    className?: string
+}
+
+export function CategoryLinks({ data, className }: CategoryLinksProps) {
+    const { id, name, categories } = data
+    const navigate = useNavigate()
+
+    return (
+        <div className={classNames(styles.container, {}, [className])}>
+            <AppLink to={RoutePath.category + id} className={styles.title}>
+                {name}
+            </AppLink>
+            <div className={styles.listContainer}>
+                {categories.map(item => {
+                    const { id, name, image } = item
+                    const categoryImage = image || productPlaceholder
+
+                    return (
+                        <div
+                            onClick={() => navigate(RoutePath.sub_category + id)}
+                            className={styles.list}
+                        >
+                            <img src={categoryImage} alt="pic" className={styles.image} />
+                            <AppLink key={id} to={RoutePath.sub_category + id}>
+                                {name}
+                            </AppLink>
+                        </div>
+                    )
+                })}
+            </div>
+        </div>
+    )
+}
